test/programytest/storage/stores/sql/store/test_learnf.py
import unittest

from programytest.storage.asserts.store.assert_learnf import LearnfStoreAsserts

from programy.storage.stores.sql.store.learnf import SQLLearnfStore
from programy.storage.stores.sql.engine import SQLStorageEngine
from programy.storage.stores.sql.config import SQLStorageConfiguration

import programytest.storage.engines as Engines


class SQLLearnfStoreTests(LearnfStoreAsserts):

    @unittest.skipIf(Engines.sql is False, Engines.sql_disabled)
    def test_initialise(self):
        config = SQLStorageConfiguration()
        engine = SQLStorageEngine(config)
        engine.initialise()
        store = SQLLearnfStore(engine)
        self.assertEqual(store.storage_engine, engine)

    @unittest.skipIf(Engines.sql is False, Engines.sql_disabled)
    def test_save_learnf(self):
        config = SQLStorageConfiguration()
        engine = SQLStorageEngine(config)
        engine.initialise()
        store = SQLLearnfStore(engine)

        self.assert_save_learnf(store)
